move GB18030 from charset to ext
